% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography.R
\name{addSamplingPeriods}
\alias{addSamplingPeriods}
\title{Add Sampling Period Variable}
\usage{
addSamplingPeriods(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{}
}
\description{
Add Sampling Period Variable
}

