% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography.R
\name{addSamplingYear}
\alias{addSamplingYear}
\title{Add Sampling Year Variable}
\usage{
addSamplingYear(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{}
}
\description{
Add Sampling Year Variable
}

