% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arrange_ggplot2.R
\name{arrange_ggplot2}
\alias{arrange_ggplot2}
\title{arrange multiple ggplots into one window}
\usage{
arrange_ggplot2(..., nrow = NULL, ncol = NULL, as.table = FALSE)
}
\arguments{
\item{nrow=NULL}{}

\item{ncol=NULL}{}

\item{as.table=FALSE}{}
}
\description{
arrange multiple ggplots into one window
}

