% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{assignSeason}
\alias{assignSeason}
\title{Assign the appropriate season for each sampling date}
\usage{
assignSeason(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{}
}
\description{
Assign the appropriate season for each sampling date
}

