% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{calculateClimateVariables}
\alias{calculateClimateVariables}
\title{Calculate climate variables from weather data}
\usage{
calculateClimateVariables(x, climate_data, calculate_dates = "TRUE",
  Dates_dataframe = NULL, first.year = 2009)
}
\arguments{
\item{x}{survey data}

\item{climate_data}{climate data}

\item{calculate_dates}{Default is \code{TRUE}. Either \code{x} is a dataframe of survey dates from which to calculate dates for climate variables or \code{x} is a list of pre-determined dates and their locations, in which case \code{calculate_dates} should be \code{FALSE}.}
}
\description{
Calculate climate variables from weather data
}

