% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_functions.R
\name{calculateDateLags}
\alias{calculateDateLags}
\title{Calculate lag dates: Previous_Survey_Date, DaysSincePrevSurvey, DaysSinceStart}
\usage{
calculateDateLags(x)
}
\arguments{
\item{x}{Dataset}
}
\description{
calculate lag dates.
}
