% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{calculateDegreeDays}
\alias{calculateDegreeDays}
\title{Calculate Growing Degree Days}
\usage{
calculateDegreeDays(climate_data, DegreeDay_list)
}
\arguments{
\item{climate_data}{Climate dataset}

\item{DegreeDay_list}{list of separate Degree Day files}
}
\description{
I used this website to calculate growing degree days: http://www.ipm.ucdavis.edu/WEATHER/ddretrievetext.html. This function merges them together with the \code{climate_data} dataframe.
}
