% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{calculateFruitPresence}
\alias{calculateFruitPresence}
\title{Determine Fruit and Flower Presence}
\usage{
calculateFruitPresence(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{}
}
\description{
Create separate variables indicating whether fruit and fruit and flowers were present.
}

