% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_functions.R
\name{calculateInsectLags2}
\alias{calculateInsectLags2}
\title{Calculate insect presence at time lag 2}
\usage{
calculateInsectLags2(x = x, arrange.variable, grouping.variable)
}
\arguments{
\item{x}{Dataset}

\item{arrange.variable}{is variable that arranges the data}

\item{grouping.variable}{is the variable that groups the data}
}
\description{
calculate lag insects.
}
