% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_functions.R
\name{calculateLagGroupedDF}
\alias{calculateLagGroupedDF}
\title{Calculate lags on a grouped dataframe}
\usage{
calculateLagGroupedDF(x, vars, arrange.variable = "Date",
  grouping.variable = "PlantID")
}
\arguments{
\item{x}{the dataset}

\item{vars}{the vector of variables}

\item{arrange.variable}{is variable that arranges the data. Defaults to "Date".}

\item{grouping.variable}{is the variable that groups the data. Defaults to "PlantID".}
}
\description{
the function first arranges by Date
}
