% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_functions.R
\name{calculateSizeLags}
\alias{calculateSizeLags}
\title{Calculate size variables at the previous time step}
\usage{
calculateSizeLags(x = x, arrange.variable, grouping.variable)
}
\arguments{
\item{x}{Dataset}

\item{arrange.variable}{is variable that arranges the data}

\item{grouping.variable}{is the variable that groups the data}
}
\description{
calculate lag size and fruit.
}

