% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalPvals.R
\name{categoricalPvals}
\alias{categoricalPvals}
\title{function gives chi-square and p values of all pairwise comparisons of categorical variables}
\usage{
categoricalPvals(indat)
}
\description{
user can input any dataset, and this will return only associations between categorical vars/factors
}
