% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_weather_names_function.R
\name{change_weather_names_function2}
\alias{change_weather_names_function2}
\title{Abbreviate weather variable names to two characters}
\usage{
change_weather_names_function2(y)
}
\arguments{
\item{y}{Dataset}
}
\description{
Abbreviate weather variable names to two characters
}
