% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/change_weather_names_function3.R
\name{change_weather_names_function3}
\alias{change_weather_names_function3}
\title{Abbreviate weather variable names}
\usage{
change_weather_names_function3(y)
}
\arguments{
\item{y}{}
}
\description{
Abbreviate weather variable names
}

