% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlotSurveys.R
\name{changeVariableNames}
\alias{changeVariableNames}
\title{Abbreviate Variable Names}
\usage{
changeVariableNames(Plot_Surveys)
}
\description{
Steps:
\itemize{
 	\item Replace "CACA_on_Pusilla" with "P_Ca"
 	\item Replace "MEPR_on_Pusilla" with "P_Me"
 	\item Replace "CHVI_on_Pusilla" with "P_Ch"
 	\item Replace "UnknwnMoth_on_Pusilla" with "P_Umoth"
 	\item Replace "Old_Moth_Evidence_Pusilla" with "P_Omoth"
 	\item Replace "Height_Pusilla" with "P_H"
 	\item Replace "Percent_Cover_Pusilla" with "P_PC"
 	\item Replace "CACA_on_Stricta" with "S_Ca"
 	\item Replace "MEPR_on_Stricta" with "S_Me"
 	\item Replace "CHVI_on_Stricta" with "S_Ch"
 	\item Replace "UnknwnMoth_on_Stricta" with "S_Umoth"
 	\item Replace "Old_Moth_Evidence_Stricta" with "S_Omoth"
 	\item Replace "Height_Stricta" with "S_H"
 	\item Replace "Percent_Cover_Stricta" with "S_PC"
	}
}

