% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{Complete_Surveys_function}
\alias{Complete_Surveys_function}
\title{Determine if a survey was complete}
\usage{
Complete_Surveys_function(x)
}
\arguments{
\item{x}{}
}
\description{
Determine if a survey was complete
}

