% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confirmed_Dead_Missing_Function.R
\name{Confirmed_Dead_Missing_Function}
\alias{Confirmed_Dead_Missing_Function}
\title{A plant is confirmed dead/missing if there area at least 2 observations (> 1) of the plant being dead and/or missing.}
\usage{
Confirmed_Dead_Missing_Function(x)
}
\arguments{
\item{x}{Vector of {0,1} values.}
}
\description{
A plant is confirmed dead/missing if there area at least 2 observations (> 1) of the plant being dead and/or missing.
}

