% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuousPvals.R
\name{continuousPvals}
\alias{continuousPvals}
\title{function gives correlations and p values of all pairwise comparisons of continuous variables}
\usage{
continuousPvals(indat)
}
\description{
user can input any dataset, and this will return only associations between continuous variables
}
