% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_Magic_Function.R
\name{Convert_Magic_Function}
\alias{Convert_Magic_Function}
\title{Convert the format of multiple columns}
\usage{
Convert_Magic_Function(obj, type)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Convert the format of multiple columns
}
\references{
found function here on 9sep13: http://stackoverflow.com/questions/11261399/function-for-converting-dataframe-column-type
}

