% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_data_sets.R
\name{createFallWinterSurveysDataset}
\alias{createFallWinterSurveysDataset}
\title{Create Dataset with Fall/Winter Surveys}
\usage{
createFallWinterSurveysDataset(timeseries)
}
\arguments{
\item{timeseries}{}
}
\description{
Create dataset from fall/winter surveys and calculate:
\itemize{
 \item lagged fruit values
 \item lagged dates
 \item lagged insect presence
 \item relative growth rate
}
}

