% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_sets.R
\name{createFruitYearDataset}
\alias{createFruitYearDataset}
\title{Create Dataset with Yearly Fruit Surveys}
\usage{
createFruitYearDataset(timeseries)
}
\arguments{
\item{timeseries}{Dataset}
}
\description{
Create a dataset with yearly observations of fruit, size, and insect observations. Each starts on the first day of spring of that calendar year, then ends on the last day of winter in the next calendar year (e.g., Spring 2009 - Winter 2010). Variables that are calculated included the maximum number of fruit observed, insect presence/absence during the year, and the maximum and minimum plant size that year. All years, except for 2009, have two observations per year.
The dataset is created according to the following steps:
\itemize{
 \item create a year variable (called "ObsYear")
 \item determine if insect was ever observed during the ObsYear
 \item calculate maximum and minimum plant size and volume
 \item determine if surveys were complete for the entire year (no missing information)
}
}

