% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{createInsectFactorVariables}
\alias{createInsectFactorVariables}
\title{Make Insect Factor Variables}
\usage{
createInsectFactorVariables(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{Plant Survey Dataset}
}
\description{
Make new insect variables with "yes"/"no" values instead of 0/1.
}

