% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateKML_Function.R
\name{CreateKML_Function}
\alias{CreateKML_Function}
\title{Create a KML from a dataframe}
\usage{
CreateKML_Function(dataset = dataset, proj4stringdata = proj4stringdata,
  CRSdata = CRSdata, NameField = NameField, filename = filename)
}
\arguments{
\item{dataset}{Dataset}

\item{proj4stringdata}{The projection data for the dataset.}

\item{CRSdata}{The projection to which the dataset should be converted.}

\item{NameField}{The attribute to use to name the points.}

\item{filename}{Name of the KML file to be generated.}
}
\description{
Create a KML Google Earth file from a dataframe.
}
\examples{
Plot_Info_Islands = read.csv(file = "../../../Written-Documents/Data-forms/island-data-collection-form/IslandRoadway1_MasterPlotList_OrderedbyGeography_Unsurveyed.csv", header = TRUE, stringsAsFactors = FALSE)
# Roadway 1
CreateKML_Function(dataset=Unsurveyed.Plots, proj4stringdata="+proj=utm +zone=17 +datum=WGS84", CRSdata="+proj=longlat +datum=WGS84", NameField="Tag_Number", filename="Unsurveyed.Plots")
# EXAMPLE
# create KML of unsurveyed plots - dataset "Unsurveyed"
# The points should be named based on the column "Tag_Number"
# give the KML file the name "unsurveyed_7sept14"
CreateKML(dataset=Unsurveyed, proj4stringdata="+proj=utm +zone=17 +datum=WGS84",
CRSdata="+proj=longlat +datum=WGS84", NameField="Tag_Number", 
filename="unsurveyed_7sept14")
}
\references{
# inspired by the answer from rcs on http://stackoverflow.com/questions/7813141/how-to-create-a-kml-file-using-r
}

