% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsectFunctions.R
\name{createNewInsectVariables}
\alias{createNewInsectVariables}
\title{Calculate Insect Variables}
\usage{
createNewInsectVariables(x, arrange.variable = "Date",
  grouping.variable = "PlantID")
}
\arguments{
\item{x}{the dataset}

\item{arrange.variable}{is variable that arranges the data. Defaults to "Date".}

\item{grouping.variable}{is the variable that groups the data. Defaults to "PlantID".}
}
\description{
Calculate insects at time t ("Insect_t") and native insects at time t ("NatInsect_t").
}

