% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_data_sets.R
\name{createObsYear}
\alias{createObsYear}
\title{Create Create a Fecundity Year Variable}
\usage{
createObsYear(timeseries)
}
\arguments{
\item{timeseries}{}
}
\description{
Each starts on the first day of spring of that calendar year, then ends on the last day of winter in the next calendar year (e.g., Spring 2009 - Winter 2010).
The years are defined as follows:
\itemize{
 \item Year 1 - Spring, Summer, Fall, Winter 2009, Winter 2010
 \item Year 2 - Spring, Summer, Fall, Winter 2010, Winter 2011
 \item Year 3 - Spring, Summer, Fall, Winter 2011, Winter 2012
 \item Year 4 - Spring, Summer, Fall, Winter 2012, Winter 2013
	\item Year 5 - Spring, Summer, Fall, Winter 2013, Winter 2014
}
}

