% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_display.R
\name{createOrientationArrow}
\alias{createOrientationArrow}
\title{Create an Orientation Arrow}
\usage{
createOrientationArrow(scaleBar, length, distance = 1,
  dist.units = "km")
}
\arguments{
\item{scaleBar}{result of createScaleBar()}

\item{length}{desired length of the arrow}

\item{distance}{distance between legend rectangles and the bottom of the arrow}

\item{dist.units}{units of distance "km" (kilometers) (default), "nm" (nautical miles), "mi" (statute miles)}
}
\description{
Returns a list containing:
res : coordinates to draw an arrow
coordinates of the middle of the arrow (where the "N" will be plotted)
}
