% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlotSurveys.R
\name{createPlotSurveysfromOrigPlantSurveys}
\alias{createPlotSurveysfromOrigPlantSurveys}
\title{Create occupancy plot survey data from original plant survey data}
\usage{
createPlotSurveysfromOrigPlantSurveys(Original_Plant_Data)
}
\description{
This data is from when I was surveying plants only, not also doing occupancy plot surveys. After IDing and surveying these plants, I deleted these IDs and renamed the plants so the IDs here DO NOT match the surveys and IDs in the demography study data.
Specific rules for filling the plot surveys:
\itemize{
 \item for these dates I am not certain that I surveyed all plants in the plots therefore either something is present or an NA
}
}

