% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlotSurveys.R
\name{createPlotSurveysfromPlantSurveys}
\alias{createPlotSurveysfromPlantSurveys}
\title{Create occupancy plot survey data from plant surveys}
\usage{
createPlotSurveysfromPlantSurveys(Plant_Surveys, Plot_Surveys, D_Plot_Surveys)
}
\description{
This data is from when I was surveying plants only, not also doing plot surveys.
Specific rules for filling the plot surveys:
\itemize{
 \item I can only add absence data when I know that I surveyed all the plants that day; this requires creating a list of possible plants alive for each survey day
 \item filter plant surveys by dates that are not in plot surveys and also not in demographic plot surveys (I don't want to replicate plot, date combos)
}
}

