% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_display.R
\name{createScaleBar}
\alias{createScaleBar}
\title{Create a Scale Bar for a ggmap}
\usage{
createScaleBar(lon, lat, distanceLon, distanceLat, distanceLegend,
  dist.units = "km")
}
\arguments{
\item{lon}{longitude of the bottom left point of the first rectangle to draw #' @param lat latitude of the bottom left point of the first rectangle to draw}

\item{distanceLon}{length of each rectangle}

\item{distanceLat}{width of each rectangle}

\item{distanceLegend}{distance between rectangles and legend texts}

\item{dist.units}{units of distance "km" (kilometers) (default), "nm" (nautical miles), "mi" (statute miles)}
}
\description{
Return a list whose elements are:
rectangle - a data.frame containing the coordinates to draw the first rectangle
rectangle2 - a data.frame containing the coordinates to draw the second rectangle
legend - a data.frame containing the coordinates of the legend texts, and the texts as well
}

