% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_data_sets.R
\name{createSpringSummerSurveysDataset}
\alias{createSpringSummerSurveysDataset}
\title{Create Dataset with Spring/Summer Surveys}
\usage{
createSpringSummerSurveysDataset(timeseries)
}
\arguments{
\item{timeseries}{}
}
\description{
Create dataset from spring/summer surveys and calculate:
\itemize{
 \item lagged fruit values
 \item lagged dates
 \item lagged insect presence
 \item relative growth rate
}
}

