% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSurvivalDataset.R
\name{createSurvivalDataset}
\alias{createSurvivalDataset}
\title{Create the Survival Dataset}
\usage{
createSurvivalDataset(timeseries)
}
\arguments{
\item{timeseries}{Dataset}
}
\description{
Steps to the creation of the dataset:
\itemize{
 \item restrict to plants that were first surveyed during the first survey of the study
 \item removed plants subject to management action, including:
\itemize{
 \item HBSPOS10
 \item HBSPOS11
 \item HBSPOS12
 \item HBSPOS15
 \item HBSPOS16
 \item HBSPOS17
 \item HBSPOS18
 \item HBSPOS5
}
 \item summarize whether plant was every infested with an insect
 \item summarize whether a plant died during the study
}
}

