% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProc.R
\name{dataProc}
\alias{dataProc}
\title{takes any dataset, checks format of each variable
Based on distribution from original variable, randomly samples into a new dataset n times}
\usage{
dataProc(inputData, n, seed, dateFormat = "\%Y\%m\%d")
}
\description{
Returns distributionally similar dataset
}
