% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlantSurveys.R
\name{determineCompleteSurveys}
\alias{determineCompleteSurveys}
\title{Determine whether a survey was complete}
\usage{
determineCompleteSurveys(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{Plant Survey Dataset}
}
\description{
Determine whether all data was collected during a survey.
Determine complete surveys for:
\itemize{
 \item insect surveys
 \item plant size
 \item size measurements (cm)
 \item fruit, flowers, and flower buds
 \item all data
	}
}

