% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlantSurveys.R
\name{determineInsectPresenceDuringStudy}
\alias{determineInsectPresenceDuringStudy}
\title{Determine if an insect species was ever detected during the study period}
\usage{
determineInsectPresenceDuringStudy(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{Plant Survey Dataset}
}
\description{
Determine if an insect species was ever detected during the study period.
}

