% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_function.R
\name{fecundity_rescale_function}
\alias{fecundity_rescale_function}
\title{Fecundity Rescale Function}
\usage{
fecundity_rescale_function(x = x)
}
\description{
Only rescale once the fecundity dataset is final (missing values have been removed)
}

