% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{filterClimateDataByQuality}
\alias{filterClimateDataByQuality}
\title{Filter data by quality}
\usage{
filterClimateDataByQuality(climate_data)
}
\arguments{
\item{climate_data}{Climate dataset}
}
\description{
Replace data with NA if it is of questionable quality.
\itemize{ 
		\item Filter data by quality
		\itemize{
			\item replace data with NA if it is of questionable quality (see \code{Quality_Flag_Function} function for details)
			\item replace blank values (quality is okay) with "Okay"
			\item Replace -999 and blanks with NAs
		}
}
}
