% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{findClosestWeatherStations}
\alias{findClosestWeatherStations}
\title{Find closest weather stations to each sampling location}
\usage{
findClosestWeatherStations(sites, climate_data, Distance = 85)
}
\arguments{
\item{sites}{List of sampling locations with x, y coordinates.}

\item{climate_data}{Climate dataset}

\item{Distance}{Radius (km) within which to look for climate stations for a particular location. Defaults to 85 kilometers.}
}
\description{
For each location, compile weather data from the closest weather stations.
}

