% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Date_Function.R
\name{Format_Date_Function}
\alias{Format_Date_Function}
\title{Format dates}
\usage{
Format_Date_Function(x)
}
\arguments{
\item{x}{Vectors of date data.}
}
\description{
The date format changes sometimes when I export the data from Access, sometimes exported as, for example, 1/28/13 0:00.
}

