% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Format_PlantIDs_Function.R
\name{Format_PlantIDs_Function}
\alias{Format_PlantIDs_Function}
\title{Format PlantID}
\usage{
Format_PlantIDs_Function(x)
}
\arguments{
\item{x}{Vectors of PlotPlantID data.}
}
\description{
Plants that spread into multiple plots have subscripts so that each PlotPlantID is associated with only one plot. This function renames Plant IDs unique to each plot (relevant only for plants found in multiple plots).
}

