% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{formatasFactors}
\alias{formatasFactors}
\title{Format as Factors}
\usage{
formatasFactors(x)
}
\arguments{
\item{Plant_Surveys}{Plant Survey Dataset}
}
\description{
Format variables as factors.
}

