% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weather_function.R
\name{formatClimateDataYearDayofYear}
\alias{formatClimateDataYearDayofYear}
\title{Create Year and Day of Year Variables for Climate Data}
\usage{
formatClimateDataYearDayofYear(climate_data)
}
\arguments{
\item{climate_data}{}
}
\description{
Create Year and Day of Year variables.
}

