% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{formatconvertClimateData}
\alias{formatconvertClimateData}
\title{Format and convert weather data}
\usage{
formatconvertClimateData(climate_data)
}
\arguments{
\item{climate_data}{Climate dataset}
}
\description{
For each location, compile weather data from the closest weather stations.
\itemize{
	\item Format/convert weather data
	\itemize{
		\item convert tenths of Celcius to Celsius
		\item convert tenths of Celcius to Celsius
		\item convert PRCP (in tenths of mm) to cm
		\item replace NA for precip less than 0
	}
}
}
