% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{formatWeatherStationInfo}
\alias{formatWeatherStationInfo}
\title{Format Weather Station Info}
\usage{
formatWeatherStationInfo(wstations)
}
\arguments{
\item{wstations}{List of weather stations, downloaded from the NOAA NCDC site}
}
\description{
Format Weather Station Info
}

