% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSeason.R
\name{getSeason2012}
\alias{getSeason2012}
\title{Assign season for 2012 dates}
\usage{
getSeason2012(DATES)
}
\arguments{
\item{DATES}{}
}
\description{
Assign seasons for the 2012 calendar year. Function modified from http://stackoverflow.com/questions/9500114/find-which-season-a-particular-date-belongs-to. I found solstice and equinox dates here: http://en.wikipedia.org/wiki/Solstice".
}

