% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbiplot2.R
\name{ggbiplot2}
\alias{ggbiplot2}
\title{PCA Plotting Function}
\usage{
ggbiplot2(pcobj, choices = 1:2, scale = 1, pc.biplot = TRUE,
  obs.scale = 1 - scale, var.scale = scale, groups = NULL,
  ellipse = FALSE, ellipse.prob = 0.68, labels = NULL,
  labels.size = 3, alpha = 1, var.axes = TRUE, circle = FALSE,
  circle.prob = 0.69, varname.size = 3, varname.adjust = 1.5,
  varname.abbrev = FALSE, color = muted("red"), linetype = "solid",
  alpha_arrow = 1, unit.circle.color = muted("white"),
  varname.color = "darkred", point.size = 1, line.size = line.size,
  scale_shape_manual = scale_shape_manual,
  scale_linetype_manual = scale_linetype_manual, axis.text.size = 15,
  axis.title.size = 20, xlim = xlim, ylim = ylim, border.size = 2,
  point.color = "black", pct_x_buffer = 0.5, pct_y_buffer = 0.5)
}
\arguments{
\item{pcobj}{PCA objects?}

\item{choices}{Defaults to \code{1:2}}

\item{scale}{Defaults to \code{1}}

\item{pc.biplot}{Defaults to \code{TRUE}}

\item{obs.scale}{Defaults to \code{1 - scale}}

\item{var.scale}{Defaults to \code{scale}}

\item{groups}{Defaults to \code{NULL}}

\item{ellipse}{Defaults to \code{FALSE}}

\item{ellipse.prob}{Defaults to \code{0.68}}

\item{labels}{Defaults to \code{NULL}}

\item{labels.size}{Defaults to \code{3}}

\item{alpha}{Defaults to \code{1}}

\item{var.axes}{Defaults to \code{TRUE}}

\item{circle}{Defaults to \code{FALSE}}

\item{circle.prob}{Defaults to \code{0.69}}

\item{varname.size}{Defaults to \code{3}}

\item{varname.adjust}{Defaults to \code{1.5}}

\item{varname.abbrev}{Defaults to \code{FALSE}}

\item{color}{Defaults to \code{muted("red")} # <- add new arguments to the function}

\item{linetype}{Defaults to \code{"solid"}}

\item{alpha_arrow}{Defaults to \code{1}}

\item{unit.circle.color}{Defaults to \code{muted("white")}}

\item{varname.color}{Defaults to \code{"darkred"}}

\item{point.size}{Defaults to \code{1}}

\item{line.size}{Defaults to \code{line.size}}

\item{scale_shape_manual}{Defaults to \code{scale_shape_manual}}

\item{scale_linetype_manual}{Defaults to \code{scale_linetype_manual}}

\item{axis.text.size}{Defaults to \code{15}}

\item{axis.title.size}{Defaults to \code{20}}

\item{xlim}{Defaults to \code{xlim}}

\item{ylim}{Defaults to \code{ylim}}

\item{border.size}{Defaults to \code{2}}

\item{point.color}{Defaults to \code{"black"}}

\item{pct_x_buffer}{Defaults to \code{0.5}}

\item{pct_y_buffer}{Defaults to \code{0.5}
Function to control graphics when plotting PCA results produced in R.}

\item{DATES}{I don't remember what this is}
}
\description{
PCA Plotting Function
}
