% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{keepUSDAData}
\alias{keepUSDAData}
\title{Keep USDA data}
\usage{
keepUSDAData(timeseries)
}
\arguments{
\item{timeseries}{Dataset of plant surveys}
}
\description{
Save data on USDA managed plants including observations of the following plants after 2010-10-01 (first observed management by USDA on 2011-05-16):
\itemize{
		\item HBSPOS10
		\item HBSPOS11
		\item HBSPOS12
		\item HBSPOS15
		\item HBSPOS16
		\item HBSPOS17
		\item HBSPOS18
		\item HBSPOS5
}
}

