% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KES_ggplot_theme.R
\name{KEStheme}
\alias{KEStheme}
\title{Publication-ready ggplot2 Theme}
\usage{
KEStheme(base_size = 10, base_family = "")
}
\arguments{
\item{base_size}{Base font size. Defaults to 10.}

\item{base_family}{Defaults to default.}
}
\description{
My personal ggplot2 theme.
}
