% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_functions.R
\name{lag_insects_function}
\alias{lag_insects_function}
\title{Calculate insect presence at previous time step}
\usage{
lag_insects_function(x = x)
}
\arguments{
\item{x}{Dataset}
}
\description{
calculate lag insects.
}

