% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lag_insects_yr_function.R
\name{lag_insects_yr_function}
\alias{lag_insects_yr_function}
\title{Calculate insect presence during the previous year}
\usage{
lag_insects_yr_function(x = x)
}
\arguments{
\item{x}{}
}
\description{
calculate lag insects during the previous year.
}

