% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_functions1.R
\name{lag_size_fruit_function_GTMNERR}
\alias{lag_size_fruit_function_GTMNERR}
\title{Calculate size and fruit production variables at the previous time step}
\usage{
lag_size_fruit_function_GTMNERR(x = x)
}
\arguments{
\item{x}{Dataset}
}
\description{
calculate lag size and fruit.
}

