% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Max.R
\name{Maximum}
\alias{Maximum}
\title{Return the maximum value of a vector, after removing NAs.}
\usage{
Maximum(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Written for use in the tables::tabular function to create publication-ready tables.
}

