% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{Maximum2}
\alias{Maximum2}
\title{Return the maximum value of a vector, after removing NAs. If all NAs, return a zero.}
\usage{
Maximum2(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Return the maximum value of a vector, after removing NAs. If all NAs, return a zero.
}

