% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{mergeClimateFiles}
\alias{mergeClimateFiles}
\title{Merge Weather Data Files and Format Column Names}
\usage{
mergeClimateFiles(climate_file_names)
}
\arguments{
\item{climate_file_names}{Climate dataset}
}
\description{
For each variable, NOAA uses generic column names "Measurement.Flag", "Quality.Flag", "Source.Flag", "Time.of.Observation" to ensure that the appropriate columns are merged together. This function renames these columns by pasting the name with the name of the variable to which it refers e.g., the "Measurement.Flag" column directly after "PRCP" will become "PRCP.Measurement.Flag" for each weather variable, take weather variable name and paste it to the names of the 4 following columns.
}
