% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlotSurveys.R
\name{mergeDuplicateSurveys}
\alias{mergeDuplicateSurveys}
\title{Merge Duplicate Surveys}
\usage{
mergeDuplicateSurveys(E)
}
\description{
Merge duplicate surveys (two or more surveys on the same date) and add columns for CACA presence and MEPR presence (regardless of which cactus the insects were found).
}

