% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePlantRecordsfromMultiplePlots.R
\name{mergePlantRecordsfromMultiplePlots}
\alias{mergePlantRecordsfromMultiplePlots}
\title{Merge Plants from Multiple Plots}
\usage{
mergePlantRecordsfromMultiplePlots(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{Plant survey dataset}
}
\description{
Only merge data collected on the same day. Merge survey data for individual plants in more than one plot. Then combine back into one file with surveys of plants in only one plot.
}

