% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography.R
\name{mergePlantSurveysPlantInfo}
\alias{mergePlantSurveysPlantInfo}
\title{Merge Plant Info and Plant Surveys data files}
\usage{
mergePlantSurveysPlantInfo(Plant_Surveys, Plant_Info)
}
\arguments{
\item{Plant_Surveys}{}

\item{Plant_Info}{}
}
\description{
Add Island, Cluster, and Host Species information to the Plant_Surveys dataset.
}

