% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlotSurveys.R
\name{mergePlantSurveysPlotSurveys}
\alias{mergePlantSurveysPlotSurveys}
\title{fill occupancy plot surveys with info from plant surveys}
\usage{
mergePlantSurveysPlotSurveys(Plot_Surveys, Plant_Surveys)
}
\description{
This data is from when I was specifically doing occupancy plot surveys.
Specific rules for filling the plot surveys:
\itemize{
 \item percent cover: get from plot survey data; if no plot survey data for that tag number and date, fill with NA
 \item height: get from plot survey data; if it was not recorded in the plot survey then use max. plant height from the plant survey data
 \item insect presence/absence: get from plot survey data; if it was not recorded in the plot survey then use max. plant height from the plant survey data
 \item limit to plot surveys after Dec. 2, 2012 (I wasn't surveying individual plants before this)
 \item limit to Tag_Numbers that have marked plants
}
}

