% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlotSurveys.R
\name{mergePlotSurveysDemoPlotSurveys}
\alias{mergePlotSurveysDemoPlotSurveys}
\title{fill demography plot surveys with info from plant surveys}
\usage{
mergePlotSurveysDemoPlotSurveys(D_Plot_Surveys, Plant_Surveys)
}
\description{
I only entered a plot as surveyed in this dataset if I had completely marked/surveyed all plants in the plot. I entered data in this dataset near the beginning of the demography study, when I was setting the study up.
Specific rules for filling the plot surveys:
\itemize{
 \item only use demography plot surveys that are not a duplicate of the occupancy plot surveys
}
}

