% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Min.R
\name{Min}
\alias{Min}
\title{Return the minimum value of a vector, after removing NAs.}
\usage{
Min(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Written for use in the tables::tabular function to create publication-ready tables.
}

