% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{Minimum}
\alias{Minimum}
\title{Return the minimum value of a vector, after removing NAs.}
\usage{
Minimum(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Written for use in the tables::tabular function to create publication-ready tables.
}

