% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_PCA_table_function.R
\name{movetolast}
\alias{movetolast}
\title{Move certain columns to the end of a dataframe}
\usage{
movetolast(data, move)
}
\arguments{
\item{data}{Dataset}

\item{move}{Variable to move? number of places to move?}
}
\description{
Move certain columns to the end of a dataframe
}

