% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{mysum}
\alias{mysum}
\title{Return NA for vector of NAs}
\usage{
mysum(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Return NA for vector of NAs
}
\references{
found here on Nov. 19, 2014: http://tolstoy.newcastle.edu.au/R/help/02a/3218.html
}
