% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mysum1.R
\name{mysum1}
\alias{mysum1}
\title{Return NA for sums below 1}
\usage{
mysum1(x)
}
\arguments{
\item{x}{Vector of data.}
}
\description{
Return NA for sums below 1
}
\references{
inspired by: http://tolstoy.newcastle.edu.au/R/help/02a/3218.html
}

