% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{mysum2}
\alias{mysum2}
\title{Return 1 for sums greater than 0.}
\usage{
mysum2(x)
}
\arguments{
\item{x}{Vector of data.}
}
\description{
Return 1 for sums greater than 0.
}
\references{
inspired by: http://tolstoy.newcastle.edu.au/R/help/02a/3218.html
}
