% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{mysum3}
\alias{mysum3}
\title{Return 1 for sums greater than 0, as well as for when vector is all NA.}
\usage{
mysum3(x)
}
\arguments{
\item{x}{Vector of data.}
}
\description{
Return 1 for sums greater than 0, as well as for when vector is all NA.
}
\references{
inspired by: http://tolstoy.newcastle.edu.au/R/help/02a/3218.html
}

