% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yes_No_NA.R
\name{NA_Function}
\alias{NA_Function}
\title{Replace values with NA.}
\usage{
NA_Function(x)
}
\arguments{
\item{x}{Vector of data.}
}
\description{
Replace values with NA.
}
\examples{
x = c(1,2,3,4," ", "Not Recorded")
NA_Function(x)

}
