% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsCatVars.R
\name{pairsCatVars}
\alias{pairsCatVars}
\title{returns list of all possible pairwise categorical/factor variables}
\usage{
pairsCatVars(indat)
}
\description{
returns list of all possible pairwise categorical/factor variables
}
