% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsContVars.R
\name{pairsContVars}
\alias{pairsContVars}
\title{returns list of all possible pairwise continuous variables}
\usage{
pairsContVars(indat)
}
\description{
returns list of all possible pairwise continuous variables
}
