% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_filter_function.R
\name{PCA_filter_function}
\alias{PCA_filter_function}
\title{filter data to make PCA}
\usage{
PCA_filter_function(x = x, species = species)
}
\arguments{
\item{x}{Dataset}

\item{species}{The species to which the data should be restricted}
}
\description{
Filter data and get complete cases to prepare data  for PCA.
}

