% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCA_table_function.R
\name{PCA_table_function}
\alias{PCA_table_function}
\title{Build a table of PCA results.}
\usage{
PCA_table_function(x, data.type)
}
\arguments{
\item{x}{output of the \code{prcomp} function.
Label to append to PC 1 and PC 2 column headers; particularly useful if one wants to display data from multiple PCAs.}
}
\description{
Build a table of PCA results.
}

