% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{Percentage}
\alias{Percentage}
\title{Return the percentage of a vector with up to one decimal places, after removing NAs.}
\usage{
Percentage(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Written for use in the tables::tabular function to create publication-ready tables.
}

