% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcomp_PCA_table_function.R
\name{prcomp_PCA_table_function}
\alias{prcomp_PCA_table_function}
\title{Build a table of PCA results.}
\usage{
prcomp_PCA_table_function(x, data.type)
}
\arguments{
\item{x}{output of the \code{prcomp} function.}
}
\description{
Label to append to PC 1 and PC 2 column headers; particularly useful if one wants to display data from multiple PCAs.
}

