% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDataTransitionMatrix.R
\name{prepDataTransitionMatrix}
\alias{prepDataTransitionMatrix}
\title{Prepare Data For Transition Matrix Function}
\usage{
prepDataTransitionMatrix(Dat, SizeClass, TransitionYear, SeedSurvival,
  SeedBankSize, SeedsPerFruit)
}
\arguments{
\item{Dat}{Dataset}

\item{SizeClass}{Classes into which to divide individuals based on size}

\item{TransitionYear}{Year}

\item{SeedSurvival}{Seed survival rate}

\item{SeedBankSize}{Number of seeds in the seed bank}

\item{SeedsPerFruit}{Number of seeds produced per fruit}
}
\description{
Prepare Data For Transition Matrix Function
}

