% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_precip_PCA_data.R
\name{process_SSprecip_PCA_data}
\alias{process_SSprecip_PCA_data}
\title{Process Spring/Summer PCA Precipitation data before merging with survey data.}
\usage{
process_SSprecip_PCA_data(data)
}
\arguments{
\item{x}{Output PCA data (from SAS).}
}
\description{
For use with SAS data.
}

