% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_temp_PCA_data.R
\name{process_temp_PCA_data}
\alias{process_temp_PCA_data}
\title{Process PCA Temperature data before merging with survey data.}
\usage{
process_temp_PCA_data(data)
}
\arguments{
\item{x}{Output PCA data (from SAS).}
}
\description{
For use with SAS data.
}

