% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processData}
\alias{processData}
\title{Process dataset}
\usage{
processData(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\description{
Process data:
\itemize{
 \item rename locations, variables, and species
 \item format variables
 \item process date and check for duplicate entries
 \item create "Visit Number" field
 \item create "Fruit Presence" ariable
 \item Classify sites as coastal or inland
 \item Create variables for insect presence during the entire study and during the current study
 \item Create season variable
 \item Classify surveys as complete
 \item calculate plant volume
	\item create PlantID2
}
}

