% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlotSurveys.R
\name{processOccupancyPlotInfo}
\alias{processOccupancyPlotInfo}
\title{Process Plot Survey Data}
\usage{
processOccupancyPlotInfo(Plot_Info)
}
\description{
Steps:
\itemize{
 \item fix Cluster variable name
 \item include "In Demography Study" information
 \item add vegetation information
 \item add elevation, distance to water information
 \item limit to only plots still in study
 \item make coordinates numeric
	}
}

