% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlotSurveys.R
\name{processOccupancyPlotSurveys}
\alias{processOccupancyPlotSurveys}
\title{Process Plot Survey Data}
\usage{
processOccupancyPlotSurveys(Plot_Surveys)
}
\description{
Steps:
\itemize{
 \item limit to only plots still in study
 \item format date
 \item Change Yes/No to 1/0
 \item make variables numeric
 \item If pusilla is not present, change all pusilla-dependent data to NA
 \item If stricta is not present, change all stricta-dependent data to NA
	}
}

