% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPlantSurveys.R
\name{processPlantSurveys}
\alias{processPlantSurveys}
\title{Process Plant Survey Data}
\usage{
processPlantSurveys(Plant_Surveys)
}
\description{
Steps:
\itemize{
 \item Fix column names
 \item format dates
 \item format PlantIDs
 \item format convert "999" values to NA
 \item format change "yes"/"no" values of the insect survey, missing, and dead columns to 0/1
 \item format numeric columns
 \item add total segment column
 \item add fruit count column
 \item add fruit/flower count column
	}
}

