% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processWeather.R
\name{processWeather}
\alias{processWeather}
\title{Process Weather Data}
\usage{
processWeather()
}
\arguments{
\item{timeseries}{}
}
\description{
Steps:
\itemize{
 \item process weather station (\code{wstations}) list
 \item for individual weather station files:
		\itemize{
			\item for each variable, NOAA uses generic column names: "Measurement.Flag", "Quality.Flag", "Source.Flag", "Time.of.Observation"
			\item to ensure that the appropriate columns are merged together, I will rename these columns by pasting the name with the name of the variable to which it refers
			\item e.g., the "Measurement.Flag" column directly after "PRCP" will become "PRCP.Measurement.Flag"
		}
 \item Filter data by quality
		\itemize{
			\item replace data with NA if it is of questionable quality (see \code{Quality_Flag_Function} function for details)
			\item replace blank values (quality is okay) with "Okay"
			\item Replace -999 and blanks with NAs
		}
 \item Format/convert weather data
		\itemize{
			\item convert tenths of Celcius to Celsius
			\item convert tenths of Celcius to Celsius
			\item convert PRCP (in tenths of mm) to cm
			\item replace NA for precip less than 0
		}
 \item Restrict start and end dates
 \item Find closest weather stations to each sampling location
 \item Choose closest weather variable measurement for each Location/Date combo
 \item Format Time variables
 \item calculate growing degree days
		\itemize{
			\item used this website to calculate growing degree days: http://www.ipm.ucdavis.edu/WEATHER/ddretrievetext.html
		}
 \item save data
}
}

