% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{Quality_Flag_Function}
\alias{Quality_Flag_Function}
\title{Filter Data by Quality Flag}
\usage{
Quality_Flag_Function(x, y)
}
\arguments{
\item{x}{Dataset with which to evaluate quality}

\item{y}{Dataset containing quality information}
}
\description{
If the quality flag from a NOAA dataset is G, I, K, L, N, O, then change the weather value to "NA"
Table 2 (Quality Flag/Attribute) from the NOAA documentation:
\itemize{
 \item Blank = did not fail any quality assurance check D = failed duplicate check
 \item G = failed gap check
 \item I = failed internal consistency check
 \item K = failed streak/frequent-value check
 \item L = failed check on length of multiday period M = failed mega-consistency check
 \item N = failed naught check
 \item O = failed climatological outlier check
	}
}

