% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeData.R
\name{removeData}
\alias{removeData}
\title{Remove Data}
\usage{
removeData(timeseries)
}
\arguments{
\item{timeseries}{Dataset of plant surveys}
}
\description{
Remove from the dataset:
\itemize{
 \item any O. engelmannii
 \item observations of the following plants after 2011-05-16:
		\itemize{
			\item HBSPOS10
			\item HBSPOS11
			\item HBSPOS12
			\item HBSPOS15
			\item HBSPOS16
			\item HBSPOS17
			\item HBSPOS18
			\item HBSPOS5
		}
 \item all observations of the following plants:
		\itemize{
			\item MBOS3
			\item BLSPOH1
			\item BLSPOH3
			\item NOH1A
		}
 \item duplicate observations of dead plants
}
}

