% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_function.R
\name{renameLocations}
\alias{renameLocations}
\title{Rename Locations}
\usage{
renameLocations(dat)
}
\arguments{
\item{dat}{Dataframe}
}
\description{
Rename Mexico Beach, Nokuse, and Sweetwater in the dataset.
}

