% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlantSurveys.R
\name{renameSpecies}
\alias{renameSpecies}
\title{Rename Species levels}
\usage{
renameSpecies(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{Plant Survey Dataset}
}
\description{
Rename Species levels
}

