% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Replace_0_w_No_Function.R
\name{Replace_0_w_No_Function}
\alias{Replace_0_w_No_Function}
\title{Replace 0 with "No".}
\usage{
Replace_0_w_No_Function(x)
}
\arguments{
\item{x}{Vector of data.}
}
\description{
Replace 0 with "No".
}

