% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Replace_1_w_Yes_Function.R
\name{Replace_1_w_Yes_Function}
\alias{Replace_1_w_Yes_Function}
\title{Replace 1 with "Yes".}
\usage{
Replace_1_w_Yes_Function(x)
}
\arguments{
\item{x}{Vector of data.}
}
\description{
Replace 1 with "Yes".
}

