% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{Replace_Blank_w_Okay_Function}
\alias{Replace_Blank_w_Okay_Function}
\title{Replace blank values (quality is okay) with "Okay"}
\usage{
Replace_Blank_w_Okay_Function(x)
}
\arguments{
\item{x}{Dataset}
}
\description{
Replace blank values (quality is okay) with "Okay"
}
