% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yes_No_NA.R
\name{Replace_NA_w_Blank_Function}
\alias{Replace_NA_w_Blank_Function}
\title{Replace NA with "".}
\usage{
Replace_NA_w_Blank_Function(x)
}
\arguments{
\item{x}{Vector of data.}
}
\description{
Replace NAs with no text so they are not printed on datasheets.
}
