% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rescale_function.R
\name{rescale_function}
\alias{rescale_function}
\title{Rescale Function}
\usage{
rescale_function(x = x)
}
\description{
Only rescale once the dataset is final (missing values have been removed)
}

