% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGR_function.R
\name{RGR_function}
\alias{RGR_function}
\title{Calculate Relative Growth Rate}
\usage{
RGR_function(x = x)
}
\arguments{
\item{x}{Dataset to which RGR variables should be added.}
}
\description{
formula for relative growth rate is from Paine, C. E. T., Marthews, T. R., Vogt, D. R., Purves, D., Rees, M., Hector, A., & Turnbull, L. A. (2011). How to fit nonlinear plant growth models and calculate growth rates: an update for ecologists. Methods in Ecology and Evolution, no–no. doi:10.1111/j.2041-210X.2011.00155.x
absolute growth rate (AGR) is the derivative with respect to time of the function used to predict biomass, and RGR is simply AGR divided by the current biomass
RGR is (dM/dt)/M
CURRENT FORMULA FOR RGR
after meeting with Mary and Bob (1 Dec 2014) Bob suggested that I divide by plant size at time t-1
scale by DaysSincePrevTime
}

