% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round_df}
\alias{round_df}
\title{Round Function}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{Dataframe}

\item{digits}{Number of units to round to}
}
\description{
Round multiple columns in a dataframe
}
