% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_statistics.R
\name{safe.ifelse}
\alias{safe.ifelse}
\title{Ifelse function that retains the original class of a variable}
\usage{
safe.ifelse(cond, yes, no)
}
\arguments{
\item{cond}{Condition}

\item{yes}{action if variable meets condition}

\item{no}{action if variable does not meet condition}
}
\description{
Retain the class of a variable when using ifelse. Copied from http://stackoverflow.com/questions/6668963/how-to-prevent-ifelse-from-turning-date-objects-into-numeric-objects.
}
