% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleUnivariate.R
\name{sampleUnivariate}
\alias{sampleUnivariate}
\title{Univariate sampling of variables through a dataset}
\usage{
sampleUnivariate(inputData, n, dateFormat = "\%Y\%m\%d")
}
\arguments{
\item{inputData}{"true" dataset that you want to sample from}

\item{n}{number of rows in output/simulated dataset}

\item{dateFormat}{format of the date variable (if one exists) in the original dataset}
}
\value{
simData -- a dataset with n rows that is distributionally similar to input dataset
}
\description{
Takes any dataset, checks format of each variable
Based on distribution from original variable, randomly samples into a new dataset n times
}
\details{
Returns distributionally similar dataset
}
\examples{
sampleUnivariate()
}
