% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_PCA_table_function.R
\name{sas_PCA_table_function}
\alias{sas_PCA_table_function}
\title{Combine PCA results from SAS into a table}
\usage{
sas_PCA_table_function(eigens, factors, rfactors, dataset_type = "all",
  round_n = 3, n.axes = 2)
}
\arguments{
\item{eigens}{SAS output of eigenvalues and variance explained}

\item{factors}{SAS output of factor pattern}

\item{rfactors}{SAS output of rotated factor pattern}

\item{dataset_type}{Default value is "all"}

\item{round_n}{Default value is 3.}

\item{n.axes}{Default value is 2.}
}
\description{
Process PCA results from SAS.
}

