% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAS_PCA_table_names.R
\name{sas_PCA_table_names}
\alias{sas_PCA_table_names}
\title{Change PCA Table names}
\usage{
sas_PCA_table_names(table)
}
\arguments{
\item{table}{created with the \code{sas_prcomp_PCA_table_function} function.}
}
\description{
First create a table from SAS output, then use this function to change the values of a couple of columns.
}

