% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processPlantInfo.R
\name{saveConfirmedDeadMissing}
\alias{saveConfirmedDeadMissing}
\title{Save File with Confirmed Dead and Missing Plants}
\usage{
saveConfirmedDeadMissing(Plant_Info)
}
\arguments{
\item{Plant_Info}{}
}
\description{
write csv with dead/missing plants - remove them from the maps
}

