% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scaleBar.R
\name{scaleBar}
\alias{scaleBar}
\title{Create a Scale Bar}
\usage{
scaleBar(lon, lat, distanceLon, distanceLat, distanceLegend, dist.unit = "km",
  rec.fill = "white", rec.colour = "black", rec2.fill = "black",
  rec2.colour = "black", legend.colour = "black", legend.size = 3,
  orientation = TRUE, arrow.length = 500, arrow.distance = 300,
  arrow.North.size = 6)
}
\arguments{
\item{lon}{longitude of the bottom left point of the first rectangle to draw #'}

\item{lat}{latitude of the bottom left point of the first rectangle to draw}

\item{distanceLon}{length of each rectangle}

\item{distanceLat}{width of each rectangle}

\item{distanceLegend}{distance between rectangles and legend texts}

\item{rec.fill}{filling colour of the rectangles (default to white, and black, resp.)}

\item{rec.colour}{colour of the rectangles (default to black for both)}

\item{rec2.fill}{filling colour of the rectangles (default to white, and black, resp.)}

\item{rec2.colour}{colour of the rectangles (default to black for both)}

\item{legend.colour}{legend colour (default to black)}

\item{legend.size}{legend size (default to 3)}

\item{orientation}{(boolean) if TRUE (default), adds an orientation arrow to the plot}

\item{arrow.length}{length of the arrow (default to 500 km)}

\item{arrow.distance}{distance between the scale bar and the bottom of the arrow (default to 300 km)}

\item{arrow.North.size}{size of the "N" letter (default to 6)}

\item{dist.units}{units of distance "km" (kilometers) (default), "nm" (nautical miles), "mi" (statute miles)}
}
\description{
This function enables to draw a scale bar on a ggplot object, and optionally an orientation arrow
}

