% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stricta_rescale_function.R
\name{stricta_rescale_function}
\alias{stricta_rescale_function}
\title{Rescale Function}
\usage{
stricta_rescale_function(x = x)
}
\description{
Only rescale once the dataset is final (missing values have been removed); the rescale function is difference for stricta - it rescales current size rather than previous size due to the small size of the dataset.
}

