% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{Saves the last n characters in a string}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{The string.}

\item{n}{The number of characters at the end of the string to save.}
}
\description{
Saves the last n characters in a string. Copied from http://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r.
}
