% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_rescale_function.R
\name{survival_rescale_function}
\alias{survival_rescale_function}
\title{survival rescale function}
\usage{
survival_rescale_function(x = x)
}
\arguments{
\item{x}{x}
}
\description{
rescale survival data
}
