% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{Transform}
\alias{Transform}
\title{Calculate log and sqrt transformations}
\usage{
Transform(data, Name)
}
\arguments{
\item{data}{Vector of data}

\item{Name}{Name to give}
}
\description{
Transform variable.
}

