% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unique.R
\name{Unique}
\alias{Unique}
\title{Return unique values of a vector after removal of NA}
\usage{
Unique(x)
}
\arguments{
\item{x}{Vector of data}
}
\description{
Return unique values of a vector after removal of NA
}
