% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Var.R
\name{Var}
\alias{Var}
\title{Return the variance of a vector with up to two decimal places, after removing NAs.}
\usage{
Var(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Written for use in the tables::tabular function to create publication-ready tables.
}

