% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weather_function.R
\name{weather_function}
\alias{weather_function}
\title{Function to process weather data}
\usage{
weather_function(x = x, climate_data = climate_data)
}
\arguments{
\item{x}{}

\item{climate_data}{}
}
\description{
Function to process weather data
}

