% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WoodyTrunk_Function.R
\name{WoodyTrunk_Function}
\alias{WoodyTrunk_Function}
\title{Standardize "WoodyTrunk" values.}
\usage{
WoodyTrunk_Function(x)
}
\arguments{
\item{x}{Vector of data.}
}
\description{
Standardize "WoodyTrunk" values.
}

