% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlantSurveys.R
\name{Yes_No_from_1_0_Function}
\alias{Yes_No_from_1_0_Function}
\title{Rename Convert 1/0 to Yes/No}
\usage{
Yes_No_from_1_0_Function(x)
}
\arguments{
\item{x}{Vector of 0/1 data}
}
\description{
Rename Convert 1/0 to Yes/No
}

