% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlantSurveys.R
\name{addSamplingYear}
\alias{addSamplingYear}
\title{Add Sampling Year Variable}
\usage{
addSamplingYear(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{Plant Survey Dataset}
}
\description{
Add Sampling Year Variable
}

