% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_ggplot2.R
\name{arrange_ggplot2}
\alias{arrange_ggplot2}
\title{arrange multiple ggplots into one window}
\usage{
arrange_ggplot2(..., nrow = NULL, ncol = NULL, as.table = FALSE)
}
\arguments{
\item{nrow}{Number of rows of plots. Defaults to \codce{NULL}.}

\item{ncol}{Number of columns of plots. Defaults to \codce{NULL}.}

\item{as.table}{I don't remember what this means. Defaults to \codce{NULL}.}
}
\description{
Arrange multiple ggplot figures in one window.
}

