% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignSeason.R
\name{assignSeason}
\alias{assignSeason}
\title{Assign Season based on Date}
\usage{
assignSeason(dat, SeasonStarts)
}
\arguments{
\item{dat}{dataframe including "Date" in \code{POSIXct} format.}

\item{SeasonStarts}{SeasonStarts}
}
\description{
Assign seasons based on date.
}
\examples{
dat = data.frame(
    Date = as.POSIXct(strptime(as.Date("2011-12-01", format = "\%Y-\%m-\%d") + 
        (0:10)*30, format="\%Y-\%m-\%d"))
)
dat \%>\% assignSeason

}
