% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlantSurveys.R
\name{calculatePlantVolume}
\alias{calculatePlantVolume}
\title{Calculate Plant Volume}
\usage{
calculatePlantVolume(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{Plant_Surveys Plant Survey Dataset}
}
\description{
Calculate plant volume as a cone, cylinder, and elliptic cylinder.
}

