% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{calculatePlantVolume}
\alias{calculatePlantVolume}
\title{Calculate Plant Volume}
\usage{
calculatePlantVolume(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{}
}
\description{
Calculate plant volume as a cone, cylinder, and elliptic cylinder.
}

