% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_functions.R
\name{calculateSizeLags}
\alias{calculateSizeLags}
\title{Calculate size variables at the previous time step}
\usage{
calculateSizeLags(x = x, arrange.variable, grouping.variable,
  vars = c("Size_t", "Height_t", "Cone_t", "Cylinder_Tall_t",
  "Cylinder_t", "Elliptic_Cylinder_t", "Size_max_t", "Size_min_t",
  "Cone_max_t", "Cone_min_t", "Cylinder_Tall_max_t",
  "Cylinder_Tall_min_t"))
}
\arguments{
\item{x}{Dataset}

\item{arrange.variable}{is variable that arranges the data}

\item{grouping.variable}{is the variable that groups the data}
}
\description{
calculate lag size and fruit.
}
