% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categoricalPvals.R
\name{categoricalPvals}
\alias{categoricalPvals}
\title{Associations between each pair of categorical variables}
\usage{
categoricalPvals(indat)
}
\arguments{
\item{inputData}{dataset that you want to sample from}
}
\value{
associations p-values for each pair of categorical variables
}
\description{
This gives associations (\Chi^2 values) and p values of all pairwise comparisons of categorical variables
User can input any dataset, and this will return only associations between categorical variables
}
\examples{
categoricalPvals()
}
