% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_weather_names_function.R
\name{change_weather_names_function1}
\alias{change_weather_names_function1}
\title{Abbreviate weather variable names}
\usage{
change_weather_names_function1(y)
}
\arguments{
\item{y}{Dataset}
}
\description{
Abbreviate weather variable names
}

