% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{Complete_Surveys_function}
\alias{Complete_Surveys_function}
\title{Determine if a survey was complete}
\usage{
Complete_Surveys_function(x)
}
\arguments{
\item{x}{Plant_Surveys Plant Survey Dataset}
}
\description{
Determine if a survey was complete
}

