% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuousPvals.R
\name{continuousPvals}
\alias{continuousPvals}
\title{Correlations between each pair of continuous variables}
\usage{
continuousPvals(inputData)
}
\arguments{
\item{inputData}{dataset that you want to sample from}
}
\value{
correlations p-values for each pair of continuous variables
}
\description{
This gives correlations and p values of all pairwise comparisons of continuous variables
User can input any dataset, and this will return only associations between continuous variables
}
\examples{
continuousPvals()
}
