% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_sets.R
\name{createObsYear}
\alias{createObsYear}
\title{Create Create a Fecundity Year Variable}
\usage{
createObsYear(timeseries)
}
\arguments{
\item{timeseries}{Dataset}
}
\description{
Each starts on the first day of spring of that calendar year, then ends on the last day of winter in the next calendar year (e.g., Spring 2009 - Winter 2010).
The years are defined as follows:
\itemize{
 \item Year 2009 - Spring, Summer, Fall, Winter 2009 - 2010
 \item Year 2010 - Spring, Summer, Fall, Winter 2010 - 2011
 \item Year 2011 - Spring, Summer, Fall, Winter 2011 - 2012
 \item Year 2012 - Spring, Summer, Fall, Winter 2012 - 2013
\item Year 2013 - Spring, Summer, Fall, Winter 2013 - 2014
}
}

