% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dead_Missing_Function.R
\name{Dead_Missing_Function}
\alias{Dead_Missing_Function}
\title{A plant has been obvserved to be dead/missing if there is at least 1 observation of it being dead or missing.}
\usage{
Dead_Missing_Function(x)
}
\arguments{
\item{x}{Vector of {0,1} data.}
}
\description{
A plant has been obvserved to be dead/missing if there is at least 1 observation of it being dead or missing.
}

