% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{determineInsectPresenceDuringStudy}
\alias{determineInsectPresenceDuringStudy}
\title{Determine if an insect species was ever detected during the study period}
\usage{
determineInsectPresenceDuringStudy(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{}
}
\description{
Determine if an insect species was ever detected during the study period.
}

