% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weather_function.R
\name{fixErroneousTemps}
\alias{fixErroneousTemps}
\title{Fix Erroneous Temperature Data}
\usage{
fixErroneousTemps(climate_data)
}
\arguments{
\item{climate_data}{}
}
\description{
If a date's minimum temperature record is greater than or equal to the maximum temperature record, replace both values with NA.
}

