% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{formatasFactors}
\alias{formatasFactors}
\title{Format as Factors}
\usage{
formatasFactors(Plant_Surveys)
}
\arguments{
\item{Plant_Surveys}{}
}
\description{
Format variables as factors.
}

