% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatasFactors.R
\name{formatasFactors}
\alias{formatasFactors}
\title{Format as Factors}
\usage{
formatasFactors(x, factors)
}
\arguments{
\item{x}{Dataset}

\item{factors}{variables to be formatted as factors}
}
\description{
Format variables as factors.
}
