% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{formatClimateDataYearDayofYear}
\alias{formatClimateDataYearDayofYear}
\title{Create Year and Day of Year Variables for Climate Data}
\usage{
formatClimateDataYearDayofYear(climate_data)
}
\arguments{
\item{climate_data}{climate data}
}
\description{
Create Year and Day of Year variables.
}

