% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{formatWeatherStationInfo}
\alias{formatWeatherStationInfo}
\title{Format Weather Station Info}
\usage{
formatWeatherStationInfo(wstations, Start_Date = "2014-01-17",
  End_Date = "2008-01-20")
}
\arguments{
\item{wstations}{List of weather stations, downloaded from the NOAA NCDC site}
}
\description{
Format Weather Station Info
}
