% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWeather.R
\name{getClimateDataByLocationDate}
\alias{getClimateDataByLocationDate}
\title{Choose closest weather variable measurement for each Location/Date combo}
\usage{
getClimateDataByLocationDate(Datalist)
}
\arguments{
\item{Datalist}{Output (list format) from the \code{findClosestWeatherStations} function.}
}
\description{
For each date and location, get weather data from the closest available weather station.
}

