% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggbiplot2.R
\name{ggbiplot2}
\alias{ggbiplot2}
\title{PCA Plotting Function}
\usage{
ggbiplot2(pcobj, choices = 1:2, scale = 1, pc.biplot = TRUE,
  obs.scale = 1 - scale, var.scale = scale, groups = NULL,
  ellipse = FALSE, ellipse.prob = 0.68, labels = NULL, labels.size = 3,
  alpha = 1, var.axes = TRUE, circle = FALSE, circle.prob = 0.69,
  varname.size = 3, varname.adjust = 1.5, varname.abbrev = FALSE,
  color = muted("red"), linetype = "solid", alpha_arrow = 1,
  unit.circle.color = muted("white"), varname.color = "darkred",
  point.size = 1, line.size = line.size,
  scale_shape_manual = scale_shape_manual,
  scale_linetype_manual = scale_linetype_manual, axis.text.size = 15,
  axis.title.size = 20, xlim = xlim, ylim = ylim, border.size = 2,
  point.color = "black", pct_x_buffer = 0.5, pct_y_buffer = 0.5)
}
\arguments{
\item{pcobj}{}

\item{choices}{= 1:2}

\item{scale}{= 1}

\item{pc.biplot}{= TRUE}

\item{obs.scale}{= 1 - scale}

\item{var.scale}{= scale}

\item{groups}{= NULL}

\item{ellipse}{= FALSE}

\item{ellipse.prob}{= 0.68}

\item{labels}{= NULL}

\item{labels.size}{= 3}

\item{alpha}{= 1}

\item{var.axes}{= TRUE}

\item{circle}{= FALSE}

\item{circle.prob}{= 0.69}

\item{varname.size}{= 3}

\item{varname.adjust}{= 1.5}

\item{varname.abbrev}{= FALSE}

\item{color}{= muted("red")# <- add new arguments to the function}

\item{linetype}{= "solid"}

\item{alpha_arrow}{= 1}

\item{varname.color}{= "darkred"}

\item{scale_shape_manual}{= scale_shape_manual}

\item{scale_linetype_manual}{= scale_linetype_manual}

\item{axis.text.size}{= 15}

\item{axis.title.size}{= 20}

\item{xlim}{= xlim}

\item{ylim}{= ylim}

\item{DATES}{}

\item{unit.circle.color=muted("white")}{}

\item{point.size=1}{}

\item{line.size=line.size}{}

\item{border.size=2}{}

\item{point.color="black"}{}

\item{pct_x_buffer=0.5}{}

\item{pct_y_buffer=0.5}{}
}
\description{
PCA Plotting Function
}

