% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_functions.R
\name{lag_dates_function}
\alias{lag_dates_function}
\title{Calculate lag dates: Previous_Survey_Date, DaysSincePrevSurvey, DaysSinceStart}
\usage{
lag_dates_function(x)
}
\arguments{
\item{x}{Dataset}
}
\description{
calculate lag dates.
}

