% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lag_functions.R
\name{lag_size_function}
\alias{lag_size_function}
\title{Calculate size variables at the previous time step}
\usage{
lag_size_function(x = x)
}
\arguments{
\item{x}{}
}
\description{
calculate lag size and fruit.
}

