% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_demography.R
\name{mergePlantSurveysPlantInfo}
\alias{mergePlantSurveysPlantInfo}
\title{Merge Plant Info and Plant Surveys data files}
\usage{
mergePlantSurveysPlantInfo(Plant_Surveys, Plant_Info)
}
\arguments{
\item{Plant_Surveys}{Plant Survey Dataset}

\item{Plant_Info}{Plant Information Dataset}
}
\description{
Add Island, Cluster, and Host Species information to the Plant_Surveys dataset.
}

