% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sas_PCA_table_function.R
\name{movetolast}
\alias{movetolast}
\title{Move certain columns to the end of a dataframe}
\usage{
movetolast(data, move)
}
\arguments{
\item{data}{}

\item{move}{}
}
\description{
Move certain columns to the end of a dataframe
}

