% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsCatVars.R
\name{pairsCatVars}
\alias{pairsCatVars}
\title{Pairs of all categorical variables}
\usage{
pairsCatVars(inputData)
}
\arguments{
\item{inputData}{dataset that you want to sample from}
}
\value{
list of all pairs of categorical variables that may or may not be associated
}
\description{
Pairs of all categorical variables
}
\examples{
pairsCatVars()

}
