% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairsContVars.R
\name{pairsContVars}
\alias{pairsContVars}
\title{Pairs of all continuous variables that may be associated}
\usage{
pairsContVars(inputData)
}
\arguments{
\item{inputData}{dataset that you want investigate}
}
\value{
list of all pairs of continuous variables that may be correlated
}
\description{
Pairs of all continuous variables that may be associated
}
\examples{
pairsContVars()


}
