% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCA_filter_function.R
\name{PCA_filter_function}
\alias{PCA_filter_function}
\title{filter data to make PCA}
\usage{
PCA_filter_function(x = x, species = species,
  restriction_list = restriction_list)
}
\arguments{
\item{x}{}

\item{species}{}

\item{restriction_list}{}
}
\description{
filter data to make PCA
}

