% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCA_function.R
\name{PCA_in_R_function}
\alias{PCA_in_R_function}
\title{Run PCA in R}
\usage{
PCA_in_R_function(x = x, variable_list = variable_list)
}
\arguments{
\item{x}{}

\item{variable_list}{}
}
\description{
Run PCA in R
}

