% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlantInfo.R
\name{processPlantInfo}
\alias{processPlantInfo}
\title{Process Plant Info}
\usage{
processPlantInfo(Plant_Info, Plot_Info)
}
\arguments{
\item{Plant_Info}{Dataset with Plant Information}

\item{Plot_Info}{Dataset with Plot Information}
}
\description{
Process Plant Info
}

