% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlantSurveys.R
\name{processPlantSurveys}
\alias{processPlantSurveys}
\title{Process Plant Survey Data}
\usage{
processPlantSurveys(Plant_Surveys)
}
\description{
Steps:
\itemize{
 \item Fix column names
 \item format dates
 \item format PlantIDs
 \item format convert "999" values to NA
 \item format change "yes"/"no" values of the insect survey, missing, and dead columns to 0/1
 \item format numeric columns
 \item add total segment column
 \item add fruit count column
 \item add fruit/flower count column
 \item check for observations of 0 pads, 0 height, or 0 width; if there are observations, stop
	}
Column Names:
\itemize{
 \item PlantMeasureID Record number if Microsoft Access database
 \item First_Observer_Initials Initials of primary observer (should always be KS, for Kristen Sauby)
 \item Second_Observer_Initials Initials of secondary observer; CJP - Cory Penca; YP - Yani Paulay; KS - Kristen Sauby; JW: Juliana Welch; CW: Cedric Worman; AP: Adam Payton
 \item Date Date
 \item PlantID Unique number for the individual plant
 \item Plant_collected Were plant samples collected?
 \item Time
 \item PlantPictures identifying numbers of photos taken
 \item PlantPictures_Direction direction that the photo(s) was(were) taken
 \item CA_t Presence/absence (1/0) of Cactoblastis cactorum
 \item CACA_collected Were samples of Cactoblastis cactorum collected?
 \item CACA_quantity Number of Cactoblastis cactorum samples collected
 \item ME_t Presence/absence (1/0) of Melitara prodenialis
 \item MEPR_collected Were samples of Melitara prodenialis collected?
 \item MEPR_quantity Number of Melitara prodenialis samples collected
 \item CH_t Presence/absence (1/0) of Chelinidea vittiger
 \item CHVI_collected Were samples of Chelinidea vittiger collected?
 \item CHVI_quantity Number of Chelinidea vittiger samples collected
 \item DA_t Presence/absence (1/0) of Dactylopius species
 \item DACT_collected Were samples of Dactylopius species collected?
 \item DACT_quantity Number of Dactylopius species samples collected
 \item Unknown_Moth_t Presence/absence (1/0) of unknown moth
 \item UnknownMoth_collected Were samples of unknown moth collected?
 \item UnknownMoth_quantity Number of unknown moth samples collected
 \item Old_Moth_Evidence_t Evidence of past moth infestation
 \item Old_Moth_Evidence_recent Whether evidence of past moth infestation appears recent or old
 \item Fungus Presence/absence (1/0) of plant fungal infection
 \item Gerstaeckeria Presence/absence (1/0) of Gerstaeckeria
 \item Plant_Segments_total Number of segments
 \item Plant_Segments_w_leaves New, green segments with leaves
 \item Plant_Segments_wo_leaves Green segments without leaves
 \item Plant_Segments_woody number of woody segments/trunks; these segments are entirely brown on the outside      
 \item Height_t maximum height in cm
 \item Width_t maximum width in cm
 \item Perpen_Width width, perpendicular to max width, in cm
 \item Num_FlowerBuds Number of flower buds
 \item Num_Fruit_red Number of red fruit
 \item Num_Fruit_green Number of green fruit
 \item Num_Flowers Number of flowers
 \item Num_Fruit use this when number is recorded but distinction by color is not made
 \item Pollinators
 \item Spiders
 \item Ants
 \item Other_collected_quantity
 \item Plant_Notes
 \item Insect_Notes        
 \item Other_Notes
 \item Dead Whether the plant is observed to be dead; 0 or 1
 \item Missing
 \item OutsideOfPlot "Yes" if plant is no longer in plot
 \item PlotPlantID Unique number for the individual plant; if a plant is counted in multiple plots, a letter is appended to the plant ID here (e.g., 9606a) and then removed for analysis
 \item Size_t The sum of Plant_Segments_total, Plant_Segments_w_leaves, Plant_Segments_wo_leaves, and Plant_Segments_woody
 \item Fruit_t The sum of Num_Fruit_red, Num_Fruit_green, and Num_Fruit
 \item Fruit_Flowers_t The sum of Num_FlowerBuds, Num_Flowers, Num_Fruit_red, Num_Fruit_green, and Num_Fruit
	}
}

