% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGTMNERRPlantSurveys.R
\name{processPlantSurveysafterMergewPlantInfo}
\alias{processPlantSurveysafterMergewPlantInfo}
\title{Process Plant Survey Data}
\usage{
processPlantSurveysafterMergewPlantInfo(x)
}
\description{
Steps:
\itemize{
 \item add column, "DemographicSurvey"
\itemize{
	\item survey 1 - spring/summer 2013
	\item survey 2 - fall/winter 2013/2014
	\item survey 3 - spring/summer 2014
	\item survey 4 - winter 2015
	\item survey 5 - spring/summer 2015
}
 \item addSamplingYear
\itemize{
	\item 2012 - Date >= "2012-12-02" & Date < "2013-05-01"
	\item 2013 - Date >= "2013-05-01" & Date < "2014-05-01"
	\item 2014 - Date >= "2014-05-01" & Date < "2015-05-01"
	\item 2015 - Date >= "2015-05-01"
}
 \item merge records for the same plant from multiple plots; the data to be merged must have been collected on the same day.
 \item Create separate variables indicating whether fruit and fruit and flowers were present
 \item Calculate plant volume as a cone, cylinder, and elliptic cylinder
 \item Determine if an insect species was ever detected during the study period
 \item Rename species levels ("Opuntia stricta" instead of "stricta" and "Opuntia pusilla" instead of "pusilla")
 \item Make new insect variables with "yes"/"no" values instead of 0/1, named "C_cactorum" and "M_prodenialis"
 \item format the variables "ClusterID", "Network", "Island", "Species", "DemographicSurvey", "SamplingYear", "Year", "Season", "C_cactorum", "M_prodenialis" as factors
	}
}

