% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleUnivariate.R
\name{sampleUnivariate}
\alias{sampleUnivariate}
\title{takes any dataset, checks format of each variable
Based on distribution from original variable, randomly samples into a new dataset n times}
\usage{
sampleUnivariate(inputData, n, seed = 1, dateFormat = "\%Y\%m\%d")
}
\description{
Returns distributionally similar dataset
}
