% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveConfirmedDeadMissing.R
\name{saveConfirmedDeadMissing}
\alias{saveConfirmedDeadMissing}
\title{Save File with Confirmed Dead and Missing Plants}
\usage{
saveConfirmedDeadMissing(Plant_Info)
}
\arguments{
\item{Plant_Info}{Dataset with Plant Information}
}
\description{
write csv with dead/missing plants - remove them from the maps
}

