% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var.R
\name{Variance}
\alias{Variance}
\title{Return the sample variance of a vector, after removing NAs, and round.}
\usage{
Variance(x, round_n = 3)
}
\arguments{
\item{x}{Vectors of data.}

\item{round_n}{Number of digits to round to.}
}
\description{
Written for use in the tables::tabular function to create publication-ready tables.
}
