% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoodyTrunk_Function.R
\name{WoodyTrunk_Function}
\alias{WoodyTrunk_Function}
\title{Standardize "WoodyTrunk" values.}
\usage{
WoodyTrunk_Function(x)
}
\arguments{
\item{x}{Vector of data.
Standardize "WoodyTrunk" values.}
}
\description{
Standardize "WoodyTrunk" values.
}

