% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_demography_2.R
\name{Yes_No_from_1_0_Function}
\alias{Yes_No_from_1_0_Function}
\title{Rename Convert 1/0 to Yes/No}
\usage{
Yes_No_from_1_0_Function(x)
}
\arguments{
\item{x}{}
}
\description{
Rename Convert 1/0 to Yes/No
}

